Moved store into its own domain
