+++ script/core/store/combineReducers.js
