+++ script/core/store/gameStateStore.js
@@ -1,0 +1,4 @@
+import createStore from './createStore';
+import reducer from './reducers';
+
+export default createStore(reducer);
