+++ script/core/store/reducers/index.js
